requirements.txt
fastapi~=0.103.1
requests~=2.31.0
python-slugify~=8.0.1
pydantic~=2.4.1
uvicorn
twilio==8.10.0